Support zombie callback handler for exited scripts
Functions and lambdas defined in a script can be registered with APIs
that outlive the script, e.g. registration via Fabric event APIs. When
the script that defined those callbacks exits, the zombie callback
handler is run when those callbacks get invoked.

If there's no user-defined zombie callback handler for a script, the
default throws an IllegalStateException.