Support Pyjinn function passed to nested interface
When passing a Pyjinn function to a nested Java interface with the
syntax `Outer.Inner(func)`, promote the Pyjinn function to an
implementation of the Java interface.

Support function promotion to an interface for a caller that's a
Class<?>. E.g.
```
Runnable = JavaClass("java.lang.Runnable")
Runnable_class = type(Runnable)  # Runnable.class
r = Runnable_class(lambda: print("Hi!"))
r.run()
```

Previously, function-interface promotion worked so long as the interface
expression didn't have a dot in it, because it was treated like a static
method call on the outer interface.

Also, fix SymbolCache.NestedClassAccessor.from() so that it reuses then
static Script cache (Script.getJavaClass()) for mapping Class<?> to
JavaClass.