General improvements to handling of **kwargs
When calling a function with **kwargs, allow kwargs to be an arbitrary
expression. Also, if the kwargs dict is empty, do not append it to the
Object[] params passed to Script.Function.call(env, params) so that
**kwargs can be used for perfect forwarding of function arguments, even
to functions that don't expect kwargs, like Java methods.

Previously, the expression following ** needed to be a variable name,
and passing an empty kwargs dict to a function resulted in an extra
KeywordArgs param being added to the params passed to
Script.Function.call(env, params).