Specialize JavaString ctor to call String ctor
When JavaClass("java.lang.String") (or in Pyjinn, `str` or `type("")`)
is used as a constructor, it was special-cased to take a single param of
type Object and convert it to String via StrClass::call (implemented as
PyjObjects.toString(params[0])). This behavior interfered with the
ability of Pyjinn scripts to invoke Java's built-in String constructors.

This change introduces special-casing of
JavaClass("org.pyjinn.interpreter.JavaString") which, when used as a
constructor, tries to invoke an appropriate built-in Java String
constructor.