Support compiled ListComprehension
Fix parsing of list comprehensions with tuple assignment, e.g.
`[... for x, y in ... ]`.

Change implementation of Context.dataStack from ArrayDeque to ArrayList
to allow random access of objects on the local stack, e.g.
Instruction.AppendListAtOffset(int offset) which pops the top of the
stack and appends it to the list residing at the given offset in the
stack.